+++ mne/io/tests/test_reference.py
@@ -142,9 +142,9 @@ def test_set_eeg_reference():
     raw = read_raw_fif(fif_fname, preload=True)
     raw.info['projs'] = []
 
-    # Test setting an average reference
+    # Test setting an average reference projection
     assert_true(not _has_eeg_average_ref_proj(raw.info['projs']))
-    reref, ref_data = set_eeg_reference(raw)
+    reref, ref_data = set_eeg_reference(raw, projection=True)
     assert_true(_has_eeg_average_ref_proj(reref.info['projs']))
     assert_true(not reref.info['projs'][0]['active'])
     assert_true(ref_data is None)
@@ -156,13 +156,13 @@ def test_set_eeg_reference():
 
     # Test setting an average reference when one was already present
     with warnings.catch_warnings(record=True):
-        reref, ref_data = set_eeg_reference(raw, copy=False)
+        reref, ref_data = set_eeg_reference(raw, copy=False, projection=True)
     assert_true(ref_data is None)
 
     # Test setting an average reference on non-preloaded data
     raw_nopreload = read_raw_fif(fif_fname, preload=False)
     raw_nopreload.info['projs'] = []
-    reref, ref_data = set_eeg_reference(raw_nopreload)
+    reref, ref_data = set_eeg_reference(raw_nopreload, projection=True)
     assert_true(_has_eeg_average_ref_proj(reref.info['projs']))
     assert_true(not reref.info['projs'][0]['active'])
 
@@ -178,7 +178,7 @@ def test_set_eeg_reference():
 
     # Test moving from custom to average reference
     reref, ref_data = set_eeg_reference(raw, ['EEG 001', 'EEG 002'])
-    reref, _ = set_eeg_reference(reref)
+    reref, _ = set_eeg_reference(reref, projection=True)
     assert_true(_has_eeg_average_ref_proj(reref.info['projs']))
     assert_equal(reref.info['custom_ref_applied'], False)
 
@@ -187,11 +187,11 @@ def test_set_eeg_reference():
     reref = raw.copy()
     reref.info['custom_ref_applied'] = True
     reref.pick_types(eeg=False)  # Cause making average ref fail
-    assert_raises(ValueError, set_eeg_reference, reref)
+    assert_raises(ValueError, set_eeg_reference, reref, projection=True)
     assert_true(reref.info['custom_ref_applied'])
 
     # Test moving from average to custom reference
-    reref, ref_data = set_eeg_reference(raw)
+    reref, ref_data = set_eeg_reference(raw, projection=True)
     reref, _ = set_eeg_reference(reref, ['EEG 001', 'EEG 002'])
     assert_true(not _has_eeg_average_ref_proj(reref.info['projs']))
     assert_equal(reref.info['custom_ref_applied'], True)
@@ -199,6 +199,23 @@ def test_set_eeg_reference():
     # Test that disabling the reference does not change anything
     reref, ref_data = set_eeg_reference(raw, [])
     assert_array_equal(raw._data, reref._data)
+
+    # Test that average reference gives identical results when calculated
+    # via SSP projection (projection=True) or directly (projection=False)
+    raw.info['projs'] = []
+    reref_1, _ = set_eeg_reference(raw.copy(), projection=True)
+    reref_1.apply_proj()
+    reref_2, _ = set_eeg_reference(raw.copy(), projection=False)
+    assert_allclose(reref_1._data, reref_2._data, rtol=1e-6, atol=1e-15)
+
+    # Test average reference without projection
+    reref, ref_data = set_eeg_reference(raw.copy(), ref_channels="average",
+                                        projection=False)
+    _test_reference(raw, reref, ref_data, eeg_chans)
+
+    # projection=True only works for ref_channels='average'
+    assert_raises(ValueError, set_eeg_reference, raw, [], True, True)
+    assert_raises(ValueError, set_eeg_reference, raw, ['EEG 001'], True, True)
 
 
 @testing.requires_testing_data